Use u32 instead of u64 presuming it would provide enough number of IDs

+++ lib/src/idg/mod.rs
@@ -1,1 +1,1 @@
-pub(crate) mod u64;
+pub(crate) mod u32;
